Add CI check to kind and kubectl install tests.

+++ packages/scripts/src/helpers/test-runner/index.ts
@@ -31,7 +31,7 @@ import {
 import { buildDevDockerImage } from '../publish/utils';
 import { PublishOptions, PublishType } from '../publish/interfaces';
 import { TestTracker } from './tracker';
-import { MAX_PROJECTS_PER_BATCH, SKIP_DOCKER_BUILD_IN_E2E } from '../config';
+import { KEEP_OPEN, MAX_PROJECTS_PER_BATCH, SKIP_DOCKER_BUILD_IN_E2E } from '../config';
 
 const logger = debugLogger('ts-scripts:cmd:test');
 
@@ -213,13 +213,13 @@ async function runE2ETest(
     if (options.testPlatform === 'kubernetes') {
         try {
             const kindInstalled = await isKindInstalled();
-            if (!kindInstalled) {
+            if (!kindInstalled && !isCI) {
                 signale.error('Please install Kind before running k8s tests. https://kind.sigs.k8s.io/docs/user/quick-start');
                 process.exit(1);
             }
 
             const kubectlInstalled = await isKubectlInstalled();
-            if (!kubectlInstalled) {
+            if (!kubectlInstalled && !isCI) {
                 signale.error('Please install kubectl before running k8s tests. https://kubernetes.io/docs/tasks/tools/');
                 process.exit(1);
             }
@@ -324,7 +324,7 @@ async function runE2ETest(
         });
     }
 
-    if (options.testPlatform === 'kubernetes') {
+    if (options.testPlatform === 'kubernetes' && !KEEP_OPEN) {
         await destroyKindCluster();
     }
 }
